Refactor of osctrl-tls to organize handlers better
