+++ tls/handlers/get.go
@@ -1,0 +1,27 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/jmpsec/osctrl/utils"
+)
+
+// RootHandler to be used as health check
+func (h *HandlersTLS) RootHandler(w http.ResponseWriter, r *http.Request) {
+	// Send response
+	utils.HTTPResponse(w, "", http.StatusOK, []byte("💥"))
+}
+
+// HealthHandler for health requests
+func (h *HandlersTLS) HealthHandler(w http.ResponseWriter, r *http.Request) {
+	h.Inc(metricHealthReq)
+	// Send response
+	utils.HTTPResponse(w, "", http.StatusOK, []byte("✅"))
+	h.Inc(metricHealthOK)
+}
+
+// ErrorHandler for error requests
+func (h *HandlersTLS) ErrorHandler(w http.ResponseWriter, r *http.Request) {
+	// Send response
+	utils.HTTPResponse(w, "", http.StatusInternalServerError, []byte("uh oh..."))
+}
